+++ CHANGELOG.md
@@ -2,7 +2,8 @@ CHANGELOG
 =========
 
 # Upcoming
-
+- Added an option to ignore strict generated exceptions while destroying socket,
+  called `ignoreStrictExceptions`. Defaults to `false`. Reason on [#11](https://github.com/findhit/proxywrap/issues/11)
 
 # 0.3.7
 - Fixed destructed problem when on non-strict #7